+++ astropy/wcs/tests/test_wcs.py
@@ -13,10 +13,12 @@ from datetime import datetime
 
 import numpy as np
 from numpy.testing import (
-    assert_allclose, assert_array_almost_equal, assert_array_almost_equal_nulp)
+    assert_allclose, assert_array_almost_equal, assert_array_almost_equal_nulp,
+    assert_array_equal)
 
 from ...tests.helper import raises, catch_warnings, pytest
 from ... import wcs
+from .. import _wcs
 from ...utils.data import (
     get_pkg_data_filenames, get_pkg_data_contents, get_pkg_data_filename)
 from ...utils.misc import NumpyRNGContext
@@ -728,6 +730,8 @@ def test_no_iteration():
     assert exc.value.args[0] == "'NewWCS' object is not iterable"
 
 
+@pytest.mark.skipif('_wcs.__version__[0] < "5"',
+                    reason="TPV only works with wcslib 5.x or later")
 def test_sip_tpv_agreement():
     sip_header = get_pkg_data_contents(
         os.path.join("data", "siponly.hdr"), encoding='binary')
@@ -740,3 +744,31 @@ def test_sip_tpv_agreement():
     assert_array_almost_equal(
         w_sip.all_pix2world([w_sip.wcs.crpix], 1),
         w_tpv.all_pix2world([w_tpv.wcs.crpix], 1))
+
+    w_sip2 = wcs.WCS(w_sip.to_header())
+    w_tpv2 = wcs.WCS(w_tpv.to_header())
+
+    assert_array_almost_equal(
+        w_sip.all_pix2world([w_sip.wcs.crpix], 1),
+        w_sip2.all_pix2world([w_sip.wcs.crpix], 1))
+    assert_array_almost_equal(
+        w_tpv.all_pix2world([w_sip.wcs.crpix], 1),
+        w_tpv2.all_pix2world([w_sip.wcs.crpix], 1))
+    assert_array_almost_equal(
+        w_sip2.all_pix2world([w_sip.wcs.crpix], 1),
+        w_tpv2.all_pix2world([w_tpv.wcs.crpix], 1))
+
+
+@pytest.mark.skipif('_wcs.__version__[0] < "5"',
+                    reason="TPV only works with wcslib 5.x or later")
+def test_tpv_copy():
+    # See #3904
+
+    tpv_header = get_pkg_data_contents(
+        os.path.join("data", "tpvonly.hdr"), encoding='binary')
+
+    w_tpv = wcs.WCS(tpv_header)
+
+    ra, dec = w_tpv.wcs_pix2world([0, 100, 200], [0, -100, 200], 0)
+    assert ra[0] != ra[1] and ra[1] != ra[2]
+    assert dec[0] != dec[1] and dec[1] != dec[2]