+++ CHANGES.rst
@@ -168,11 +168,13 @@ New Features
 
 - ``astropy.wcs``
 
-  - The included version of wcslib has been upgraded to 5.6.
+  - The included version of wcslib has been upgraded to 5.8.pre.
 
     The minimum required version of wcslib in the 4.x series remains 4.24.
 
-    The minimum required version of wcslib in the 5.x series is 5.6.
+    The minimum required version of wcslib in the 5.x series is
+    5.8.pre.  Building astropy against a wcslib 5.x prior to 5.8.pre
+    will raise an ``ImportError`` when ``astropy.wcs`` is imported.
 
     The wcslib changes relevant to astropy are: